multivlm-drive-vqa/requirements.txt
```text
torch>=2.3.0
torchvision
transformers>=4.45.0
accelerate
bitsandbytes
peft

sacrebleu
rouge-score
pycocoevalcap
scikit-learn
bert-score

Pillow
tqdm
matplotlib
wordcloud